requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-filter==24.3
Pillow==10.4.0
python-decouple==3.8
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
boto3==1.34.162
